+++ core-ng/src/main/java/core/framework/impl/web/response/BeanBody.java
@@ -1,8 +1,6 @@
 package core.framework.impl.web.response;
 
 import core.framework.impl.log.filter.BytesParam;
-import core.framework.impl.validate.ValidationException;
-import core.framework.json.JSON;
 import io.undertow.io.Sender;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -22,13 +20,8 @@ public final class BeanBody implements Body {
 
     @Override
     public void send(Sender sender, ResponseHandlerContext context) {
-        try {
-            byte[] body = context.responseBeanMapper.toJSON(bean);
-            logger.debug("[response] body={}", new BytesParam(body));
-            sender.send(ByteBuffer.wrap(body));
-        } catch (ValidationException e) {
-            logger.debug("failed to validate response bean, bean={}", JSON.toJSON(bean));  // log invalid bean for troubleshooting
-            throw e;
-        }
+        byte[] body = context.responseBeanMapper.toJSON(bean);
+        logger.debug("[response] body={}", new BytesParam(body));
+        sender.send(ByteBuffer.wrap(body));
     }
 }