+++ core-ng/src/test/java/core/framework/impl/kafka/MessageListenerThreadTest.java
@@ -1,5 +1,7 @@
 package core.framework.impl.kafka;
 
+import core.framework.util.Strings;
+import org.apache.kafka.common.header.internals.RecordHeaders;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -22,5 +24,13 @@ class MessageListenerThreadTest {
     void longProcessThreshold() {
         assertThat(thread.longProcessThreshold(Duration.ofNanos(500).toNanos(), 1, 100)).isEqualTo(5);
         assertThat(thread.longProcessThreshold(Duration.ofNanos(500).toNanos(), 1, 1)).isEqualTo(500);
+    }
+
+    @Test
+    void header() {
+        var headers = new RecordHeaders();
+        headers.add("header", Strings.bytes("value"));
+        assertThat(thread.header(headers, "header")).isEqualTo("value");
+        assertThat(thread.header(headers, "nonExisted")).isNull();
     }
 }